requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
pytz>=2023.3
flask
aiofiles>=23.1.0